Moved gitignore to root directory
