+++ .gitignore
@@ -1,7 +1,7 @@
 # See http://help.github.com/ignore-files/ for more about ignoring files.
 
 # environment variables
-/src/environments/env_var.js
+/env_var.js
 
 # compiled output
 /dist
